feat(minimap): permitir ajuste manual del zoom desde el archivo minimap.py
Se introdujo la variable local `minimap_zoom` para controlar el tamaño
de los tiles en el minimapa. Ahora se puede acercar o alejar la vista
cambiando ese valor directamente en el código, sin depender del estado global.

También se mantiene el soporte de colores diferenciados para los tiles
'R', 'T', '.', '#', y 'D', con un fallback magenta para debugging visual.
